Rename a variable throughout the script
Rename `recoveries` to `dailyRecovered`. Ashok's intention is to highlight that the change is per day.

I tend to think in the "Tidy R" manner; so with a dataframe, the observation of recoveries with a corresponding date variable means that these number of people recovered on this day. If I wanted the cumulative sum of recoveries over time, I would *summarize* the data separately using `cumSum(recoveries)` to obtain a vector of the cumulative sum of recovered persons. I like to use active participles as well, rather than past tense.

Co-Authored-By: Ashok Krishnamurthy <ashokkrish@gmail.com>